Improved handling of precipitation
A new parameter "min_newsnow_layer" is introduced to determine when precipitation is added to the snowpack. This parameter can be set independently from the other layer spacing parameters controlling model layer spacing. If the surface layer is thin, and the layer to be added is thin, the top layer will grow. Only when the top layer is sufficiently thick, precipitation will be deposited in a new layer. This avoids that precipitation is added at different timesteps, depending on the layer spacing settings.